Database integration with zookeeper
